pyproject.toml
[project]
name = "recap-attack"
version = "1.0.0"
description = "RECAP: Retrieval-based Adversarial Prompt Attack on LLMs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "huggingface-hub>=0.14.0",
]

[project.urls]
Homepage = "https://github.com/rishitchugh/recap-attack"
Repository = "https://github.com/rishitchugh/recap-attack"
